rename model.common package to model.base
