+++ src/main/java/org/sbml4j/controller/NetworksApiController.java
@@ -35,8 +35,8 @@ import org.sbml4j.model.api.network.FilterOptions;
 import org.sbml4j.model.api.network.NetworkInventoryItem;
 import org.sbml4j.model.api.network.NetworkOptions;
 import org.sbml4j.model.api.network.NodeList;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphActivityType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphActivityType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphEdgeType;
 import org.sbml4j.model.flat.FlatEdge;
 import org.sbml4j.model.provenance.ProvenanceEntity;
 import org.sbml4j.model.provenance.ProvenanceGraphActivityNode;
