+++ src/main/java/org/sbml4j/controller/OverviewApiController.java
@@ -27,8 +27,8 @@ import org.sbml4j.config.OverviewNetworkConfig;
 import org.sbml4j.config.SBML4jConfig;
 import org.sbml4j.model.api.network.NetworkInventoryItem;
 import org.sbml4j.model.api.network.OverviewNetworkItem;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphActivityType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphAgentType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphActivityType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphAgentType;
 import org.sbml4j.model.flat.FlatEdge;
 import org.sbml4j.model.flat.FlatSpecies;
 import org.sbml4j.model.provenance.ProvenanceGraphAgentNode;
