+++ src/main/java/org/sbml4j/controller/PathwaysApiController.java
@@ -29,10 +29,10 @@ import org.sbml4j.api.PathwaysApi;
 import org.sbml4j.model.api.network.NetworkInventoryItem;
 import org.sbml4j.model.api.pathway.PathwayCollectionCreationItem;
 import org.sbml4j.model.api.pathway.PathwayInventoryItem;
-import org.sbml4j.model.common.GraphEnum.NetworkMappingType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphActivityType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphAgentType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.NetworkMappingType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphActivityType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphAgentType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphEdgeType;
 import org.sbml4j.model.provenance.ProvenanceGraphActivityNode;
 import org.sbml4j.model.provenance.ProvenanceGraphAgentNode;
 import org.sbml4j.model.warehouse.DatabaseNode;
