+++ src/main/java/org/sbml4j/controller/SbmlApiController.java
@@ -27,10 +27,10 @@ import org.sbml.jsbml.Model;
 import org.sbml4j.Exception.ModelPersistenceException;
 import org.sbml4j.api.SbmlApi;
 import org.sbml4j.model.api.pathway.PathwayInventoryItem;
-import org.sbml4j.model.common.GraphEnum.FileNodeType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphActivityType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphAgentType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.FileNodeType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphActivityType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphAgentType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphEdgeType;
 import org.sbml4j.model.provenance.ProvenanceEntity;
 import org.sbml4j.model.provenance.ProvenanceGraphActivityNode;
 import org.sbml4j.model.provenance.ProvenanceGraphAgentNode;
