+++ src/main/java/org/sbml4j/model/provenance/ProvenanceGraphEdge.java
@@ -16,7 +16,7 @@ package org.sbml4j.model.provenance;
 import org.neo4j.ogm.annotation.EndNode;
 import org.neo4j.ogm.annotation.RelationshipEntity;
 import org.neo4j.ogm.annotation.StartNode;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphEdgeType;
 
 @RelationshipEntity(type="PROV")
 public class ProvenanceGraphEdge extends ProvenanceEntity {
