+++ src/main/java/org/sbml4j/service/NetworkMappingServiceImpl.java
@@ -28,11 +28,11 @@ import org.sbml.jsbml.CVTerm.Qualifier;
 import org.sbml4j.Exception.NetworkAlreadyExistsException;
 import org.sbml4j.Exception.NetworkMappingError;
 import org.sbml4j.config.SBML4jConfig;
-import org.sbml4j.model.common.GraphEnum.AnnotationName;
-import org.sbml4j.model.common.GraphEnum.ExternalResourceType;
-import org.sbml4j.model.common.GraphEnum.NetworkMappingType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphEdgeType;
-import org.sbml4j.model.common.GraphEnum.WarehouseGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.AnnotationName;
+import org.sbml4j.model.base.GraphEnum.ExternalResourceType;
+import org.sbml4j.model.base.GraphEnum.NetworkMappingType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.WarehouseGraphEdgeType;
 import org.sbml4j.model.flat.FlatEdge;
 import org.sbml4j.model.flat.FlatSpecies;
 import org.sbml4j.model.provenance.ProvenanceGraphActivityNode;
