+++ src/main/java/org/sbml4j/service/SBMLSimpleModelService.java
@@ -45,9 +45,9 @@ import org.sbml.jsbml.ext.qual.Transition;
 import org.sbml.jsbml.xml.XMLNode;
 import org.sbml4j.Exception.ModelPersistenceException;
 import org.sbml4j.config.SBML4jConfig;
-import org.sbml4j.model.common.GraphEnum.ExternalResourceType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphEdgeType;
-import org.sbml4j.model.common.GraphEnum.WarehouseGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.ExternalResourceType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.WarehouseGraphEdgeType;
 import org.sbml4j.model.provenance.ProvenanceEntity;
 import org.sbml4j.model.provenance.ProvenanceGraphActivityNode;
 import org.sbml4j.model.sbml.SBMLCompartment;
