+++ src/main/java/org/sbml4j/service/SimpleSBML/SBMLCompartmentService.java
@@ -2,7 +2,7 @@ package org.sbml4j.service.SimpleSBML;
 
 import java.util.List;
 
-import org.sbml4j.model.common.GraphBaseEntity;
+import org.sbml4j.model.base.GraphBaseEntity;
 import org.sbml4j.model.sbml.SBMLCompartment;
 import org.sbml4j.repository.sbml.SBMLCompartmentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
