+++ src/main/java/org/sbml4j/service/SimpleSBML/SBMLSimpleTransitionService.java
@@ -15,7 +15,7 @@ package org.sbml4j.service.SimpleSBML;
 
 import java.util.List;
 
-import org.sbml4j.model.common.GraphBaseEntity;
+import org.sbml4j.model.base.GraphBaseEntity;
 import org.sbml4j.model.queryResult.NonMetabolicPathwayReturnType;
 import org.sbml4j.model.sbml.simple.ext.qual.SBMLSimpleTransition;
 import org.sbml4j.repository.sbml.simple.ext.qual.SBMLSimpleTransitionRepository;
