+++ src/main/java/org/sbml4j/service/SimpleSBML/SBMLSpeciesService.java
@@ -19,7 +19,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.sbml4j.config.SBML4jConfig;
-import org.sbml4j.model.common.GraphBaseEntity;
+import org.sbml4j.model.base.GraphBaseEntity;
 import org.sbml4j.model.sbml.SBMLSpecies;
 import org.sbml4j.model.sbml.SBMLSpeciesGroup;
 import org.sbml4j.model.sbml.ext.sbml4j.BiomodelsQualifier;
