+++ src/main/java/org/sbml4j/service/WarehouseGraphService.java
@@ -15,7 +15,7 @@ package org.sbml4j.service;
 
 import java.util.List;
 
-import org.sbml4j.model.common.GraphEnum.WarehouseGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.WarehouseGraphEdgeType;
 import org.sbml4j.model.provenance.ProvenanceEntity;
 import org.sbml4j.model.provenance.ProvenanceGraphAgentNode;
 import org.sbml4j.model.provenance.ProvenanceGraphEdge;
