+++ src/main/java/org/sbml4j/service/networks/NetworkService.java
@@ -32,11 +32,11 @@ import org.sbml4j.Exception.NetworkDeletionException;
 import org.sbml4j.model.api.network.AnnotationItem;
 import org.sbml4j.model.api.network.FilterOptions;
 import org.sbml4j.model.api.network.NetworkInventoryItem;
-import org.sbml4j.model.common.GraphEnum.NetworkMappingType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphActivityType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphAgentType;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphEdgeType;
-import org.sbml4j.model.common.GraphEnum.WarehouseGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.NetworkMappingType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphActivityType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphAgentType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.WarehouseGraphEdgeType;
 import org.sbml4j.model.flat.FlatEdge;
 import org.sbml4j.model.flat.FlatSpecies;
 import org.sbml4j.model.provenance.ProvenanceEntity;
