+++ src/main/java/org/sbml4j/service/warehouse/MappingNodeService.java
@@ -23,7 +23,7 @@ import java.util.Set;
 import org.sbml4j.model.api.network.AnnotationItem;
 import org.sbml4j.model.api.network.FilterOptions;
 import org.sbml4j.model.api.network.NetworkOptions;
-import org.sbml4j.model.common.GraphEnum.NetworkMappingType;
+import org.sbml4j.model.base.GraphEnum.NetworkMappingType;
 import org.sbml4j.model.flat.FlatSpecies;
 import org.sbml4j.model.provenance.ProvenanceGraphAgentNode;
 import org.sbml4j.model.provenance.ProvenanceGraphEdge;
