+++ src/main/java/org/sbml4j/service/warehouse/PathwayCollectionNodeService.java
@@ -16,8 +16,8 @@ package org.sbml4j.service.warehouse;
 import java.util.UUID;
 
 import org.sbml4j.model.api.pathway.PathwayCollectionCreationItem;
-import org.sbml4j.model.common.GraphEnum.ProvenanceGraphEdgeType;
-import org.sbml4j.model.common.GraphEnum.WarehouseGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.ProvenanceGraphEdgeType;
+import org.sbml4j.model.base.GraphEnum.WarehouseGraphEdgeType;
 import org.sbml4j.model.provenance.ProvenanceEntity;
 import org.sbml4j.model.provenance.ProvenanceGraphActivityNode;
 import org.sbml4j.model.provenance.ProvenanceGraphAgentNode;
